requirements.txt
numpy~=1.26.4
pyqtgraph~=0.13.3
setproctitle~=1.3.3
pandas~=2.2.0
strip.pysondb~=1.5.1
GitPython~=3.1.41
cacheout~=0.16.0
tzlocal~=5.2
humanize~=4.9.0
bleak~=0.21.1
h5py~=3.10.0
ieee754~=0.10
boto3~=1.26.0
bullet~=2.2.0
toml~=0.10.2
py_cui~=0.1.6
pyserial~=3.5b0

